% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{align_signs_cpp}
\alias{align_signs_cpp}
\title{Sign Flip Alignment}
\usage{
align_signs_cpp(u_boot, u_obs)
}
\arguments{
\item{u_boot}{Bootstrap u matrix}

\item{u_obs}{Observed u matrix}
}
\value{
Sign-aligned u_boot
}
\description{
Aligns bootstrap u vectors to observed u by sign flipping.
}
\keyword{internal}
