% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_pls.R
\name{asca_term_table}
\alias{asca_term_table}
\title{Extract ASCA-PLS Term Table}
\usage{
asca_term_table(x, ...)
}
\arguments{
\item{x}{An \code{asca_pls_result}.}

\item{...}{Unused.}
}
\value{
A data frame with term diagnostics and test status.
}
\description{
Extract ASCA-PLS Term Table
}
