% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_pls.R
\name{brain_saliences}
\alias{brain_saliences}
\title{Extract ASCA-PLS Brain Saliences}
\usage{
brain_saliences(x, ...)
}
\arguments{
\item{x}{An \code{asca_pls_result}.}

\item{...}{Unused.}
}
\value{
A matrix of brain saliences.
}
\description{
Extract ASCA-PLS Brain Saliences
}
