% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_datamat}
\alias{check_datamat}
\title{Check Data Matrix Validity}
\usage{
check_datamat(x, name = "data matrix")
}
\arguments{
\item{x}{Matrix to check}

\item{name}{Name for error messages}
}
\value{
TRUE invisibly if valid
}
\description{
Validates a data matrix for PLS analysis.
}
\keyword{internal}
