% Generated by hand: do not edit by hand
\name{check_gui_deps}
\alias{check_gui_deps}
\title{Check GUI Dependencies}
\usage{
check_gui_deps()
}
\value{
Logical indicating whether all required GUI dependencies are installed.
}
\description{
Checks if all required packages for the Shiny GUI are installed.
}
\examples{
check_gui_deps()
}
