% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colwise_cor_cpp}
\alias{colwise_cor_cpp}
\title{Fast Column-wise Correlation}
\usage{
colwise_cor_cpp(X, Y)
}
\arguments{
\item{X}{First matrix}

\item{Y}{Second matrix}
}
\value{
Vector of correlations
}
\description{
Computes correlation between corresponding columns of two matrices.
}
\keyword{internal}
