% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_bsr_cpp}
\alias{compute_bsr_cpp}
\title{Compute Bootstrap Ratio}
\usage{
compute_bsr_cpp(u_sum, u_sq_sum, observed_u, num_boot)
}
\arguments{
\item{u_sum}{Sum of bootstrap u}

\item{u_sq_sum}{Sum of squared bootstrap u}

\item{observed_u}{Observed u}

\item{num_boot}{Number of bootstrap samples}
}
\value{
Bootstrap ratio matrix
}
\description{
Computes bootstrap ratio (u / SE) from accumulated sums.
}
\keyword{internal}
