% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_pls.R
\name{design_saliences}
\alias{design_saliences}
\title{Extract ASCA-PLS Design Saliences}
\usage{
design_saliences(x, ...)
}
\arguments{
\item{x}{An \code{asca_pls_result}.}

\item{...}{Unused.}
}
\value{
A matrix of design profiles in cell coordinates.
}
\description{
Extract ASCA-PLS Design Saliences
}
