% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_conditions.R
\name{.add_conditions_events}
\alias{.add_conditions_events}
\title{Add Conditions from Events Data Frame}
\usage{
.add_conditions_events(spec, events)
}
\description{
Parses BIDS-style events to determine condition structure.
}
\keyword{internal}
