% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_subjects.R
\name{.add_subjects_matrices}
\alias{.add_subjects_matrices}
\title{Add Subjects from Matrices}
\usage{
.add_subjects_matrices(spec, datamat_lst, groups)
}
\description{
Add Subjects from Matrices
}
\keyword{internal}
