% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{.matrix_to_neurovol}
\alias{.matrix_to_neurovol}
\title{Convert Matrix to NeuroVol}
\usage{
.matrix_to_neurovol(mat, mask)
}
\description{
Helper function to convert a salience/BSR matrix to NeuroVol using mask.
}
\keyword{internal}
