% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_core.R
\name{.pipeline_relativize_df}
\alias{.pipeline_relativize_df}
\title{Relativize path columns in a data.frame}
\usage{
.pipeline_relativize_df(df, root, cols)
}
\description{
Relativize path columns in a data.frame
}
\keyword{internal}
