% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_core.R
\name{.pipeline_resolve_df}
\alias{.pipeline_resolve_df}
\title{Resolve path columns in a data.frame}
\usage{
.pipeline_resolve_df(df, root, cols)
}
\description{
Resolve path columns in a data.frame
}
\keyword{internal}
