% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_boot.R
\name{.pls_boot_order_core}
\alias{.pls_boot_order_core}
\title{Core Bootstrap Order Generator (No bscan mapping)}
\usage{
.pls_boot_order_core(
  num_subj_lst,
  num_cond,
  num_boot,
  incl_seq = FALSE,
  boot_type = "strat"
)
}
\description{
Core Bootstrap Order Generator (No bscan mapping)
}
\keyword{internal}
