% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_boot.R
\name{.pls_boot_order_ssb}
\alias{.pls_boot_order_ssb}
\title{Core Bootstrap Order Generator for SSB designs}
\usage{
.pls_boot_order_ssb(
  num_subj_lst,
  num_cond,
  num_boot,
  bscan = NULL,
  incl_seq = FALSE,
  boot_type = "strat"
)
}
\description{
Core Bootstrap Order Generator for SSB designs
}
\keyword{internal}
