% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_boot.R
\name{.pls_islowvariability}
\alias{.pls_islowvariability}
\title{Low-variability check for bootstrap behavior vectors (MATLAB rri_islowvariability)}
\usage{
.pls_islowvariability(bodat, behavdata)
}
\description{
Low-variability check for bootstrap behavior vectors (MATLAB rri_islowvariability)
}
\keyword{internal}
