% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_perm.R
\name{.pls_perm_order_ssb}
\alias{.pls_perm_order_ssb}
\title{SSB Permutation Orders (Unequal n per condition)}
\usage{
.pls_perm_order_ssb(num_subj_lst, num_cond, num_perm)
}
\description{
SSB Permutation Orders (Unequal n per condition)
}
\keyword{internal}
