% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_run.R
\name{.process_bids_data}
\alias{.process_bids_data}
\title{Process BIDS Data}
\usage{
.process_bids_data(spec)
}
\description{
Converts raw BIDS data to matrices using the condition structure.
}
\keyword{internal}
