% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_run.R
\name{.process_trial_data}
\alias{.process_trial_data}
\title{Process Trial-Level Within-Subject Seed Data}
\usage{
.process_trial_data(spec)
}
\description{
Process Trial-Level Within-Subject Seed Data
}
\keyword{internal}
