% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_conditions.R
\name{.validate_dimensions}
\alias{.validate_dimensions}
\title{Validate Data Matrix Dimensions}
\usage{
.validate_dimensions(spec)
}
\description{
Validate Data Matrix Dimensions
}
\keyword{internal}
