% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brain.R
\name{.values_to_neurovol}
\alias{.values_to_neurovol}
\title{Convert Values Vector to NeuroVol}
\usage{
.values_to_neurovol(values, mask)
}
\arguments{
\item{values}{Numeric vector of values}

\item{mask}{A NeuroVol mask}
}
\value{
A NeuroVol
}
\description{
Helper to convert a vector of values (in mask order) to a NeuroVol.
}
\keyword{internal}
