% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_pvalue}
\alias{format_pvalue}
\title{Format P-value}
\usage{
format_pvalue(p, digits = 4)
}
\arguments{
\item{p}{P-value(s)}

\item{digits}{Number of digits}
}
\value{
Formatted string
}
\description{
Formats p-values for display.
}
\keyword{internal}
