% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_behavior.R
\name{generate_contrasts}
\alias{generate_contrasts}
\title{Generate Common Contrasts}
\usage{
generate_contrasts(num_groups, num_cond, type = "helmert")
}
\arguments{
\item{num_groups}{Number of groups}

\item{num_cond}{Number of conditions}

\item{type}{Type of contrast:
\itemize{
\item "helmert": Helmert contrasts
\item "deviation": Deviation from mean
\item "treatment": Treatment vs baseline
\item "polynomial": Polynomial trends
}}
}
\value{
Contrast matrix
}
\description{
Generates common contrast matrices for PLS analysis.
}
