% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mva_method.R
\name{get_method}
\alias{get_method}
\title{Get a Registered Method by Name}
\usage{
get_method(name)
}
\arguments{
\item{name}{Method name or alias}
}
\value{
An \code{MvaMethod} R6 instance
}
\description{
Get a Registered Method by Name
}
