% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plscli.R
\name{install_cli}
\alias{install_cli}
\title{Install \code{plsrri} Command Wrappers}
\usage{
install_cli(dest_dir = "~/.local/bin", overwrite = FALSE, commands = NULL)
}
\arguments{
\item{dest_dir}{Destination directory for installed commands.}

\item{overwrite}{Overwrite existing command wrappers if \code{TRUE}.}

\item{commands}{Optional character vector of command names to install.}
}
\value{
Invisibly returns the installed wrapper paths.
}
\description{
Install \code{plsrri} Command Wrappers
}
