% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mva_accessors.R
\name{n_components}
\alias{n_components}
\title{Get Number of Components}
\usage{
n_components(x, ...)
}
\arguments{
\item{x}{An \code{mva_result} or \code{pls_result} object}

\item{...}{Additional arguments}
}
\value{
Integer
}
\description{
Returns the number of components (latent variables, principal components,
independent components, etc.) in the result.
}
