% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{n_features}
\alias{n_features}
\title{Get Number of Voxels/Features}
\usage{
n_features(x)
}
\arguments{
\item{x}{A \code{pls_result} object}
}
\value{
Integer
}
\description{
Returns the number of voxels or features in a PLS result.
}
