% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{n_lv}
\alias{n_lv}
\title{Get Number of Latent Variables}
\usage{
n_lv(x)
}
\arguments{
\item{x}{A \code{pls_result} object}
}
\value{
Integer
}
\description{
Returns the number of latent variables in a PLS result.
}
