% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_pls_result}
\alias{new_pls_result}
\title{Create a PLS Result Object}
\usage{
new_pls_result(
  method,
  u,
  s,
  v,
  usc = NULL,
  vsc = NULL,
  datamatcorrs_lst = NULL,
  lvcorrs = NULL,
  perm_result = NULL,
  boot_result = NULL,
  splithalf_result = NULL,
  num_subj_lst = NULL,
  num_cond = NULL,
  bscan = NULL,
  stacked_designdata = NULL,
  stacked_behavdata = NULL,
  other_input = NULL,
  TBv = NULL,
  TBusc = NULL,
  TBvsc = NULL,
  is_struct = FALSE,
  mask = NULL
)
}
\arguments{
\item{method}{Integer, PLS method (1-6)}

\item{u}{Salience (brain loadings) matrix}

\item{s}{Singular values}

\item{v}{Design/behavior loadings matrix}

\item{usc}{Brain scores}

\item{vsc}{Design/behavior scores}

\item{datamatcorrs_lst}{Correlation matrices by group (behavior PLS)}

\item{lvcorrs}{Latent variable correlations (behavior PLS)}

\item{perm_result}{Permutation result object}

\item{boot_result}{Bootstrap result object}

\item{splithalf_result}{Split-half result object}

\item{num_subj_lst}{Number of subjects per group}

\item{num_cond}{Number of conditions}

\item{bscan}{Conditions used for behavior block (multiblock)}

\item{stacked_designdata}{Design contrast matrix (non-rotated methods)}

\item{stacked_behavdata}{Behavior data matrix}

\item{other_input}{List of other inputs (meancentering_type, cormode)}

\item{TBv}{Task/Behavior v stored separately (multiblock)}

\item{TBusc}{Task/Behavior brain scores (multiblock)}

\item{TBvsc}{Task/Behavior design scores (multiblock)}

\item{is_struct}{Structure PLS flag}

\item{mask}{Brain mask (optional)}
}
\value{
A \code{pls_result} object
}
\description{
Create a PLS Result Object
}
\keyword{internal}
