% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_pls_splithalf_result}
\alias{new_pls_splithalf_result}
\title{Create a Split-Half Validation Result Object}
\usage{
new_pls_splithalf_result(
  num_outer_perm,
  num_split,
  orig_ucorr = NULL,
  orig_vcorr = NULL,
  ucorr_prob = NULL,
  vcorr_prob = NULL,
  ucorr_ul = NULL,
  ucorr_ll = NULL,
  vcorr_ul = NULL,
  vcorr_ll = NULL
)
}
\arguments{
\item{num_outer_perm}{Number of outer permutations}

\item{num_split}{Number of splits}

\item{orig_ucorr}{Original u (brain) correlation}

\item{orig_vcorr}{Original v (design/behavior) correlation}

\item{ucorr_prob}{Probability for u correlation}

\item{vcorr_prob}{Probability for v correlation}

\item{ucorr_ul}{Upper limit for u correlation}

\item{ucorr_ll}{Lower limit for u correlation}

\item{vcorr_ul}{Upper limit for v correlation}

\item{vcorr_ll}{Lower limit for v correlation}
}
\value{
A \code{pls_splithalf_result} object
}
\description{
Create a Split-Half Validation Result Object
}
\keyword{internal}
