% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize_rows_cpp}
\alias{normalize_rows_cpp}
\title{Fast Row Normalization}
\usage{
normalize_rows_cpp(X)
}
\arguments{
\item{X}{Matrix to normalize}
}
\value{
Normalized matrix
}
\description{
Normalizes each row to unit L2 norm.
}
