% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_test_fast_cpp}
\alias{perm_test_fast_cpp}
\title{Fast Permutation Test with Pre-computed Cross-block Matrices}
\usage{
perm_test_fast_cpp(task_means_lst, permsamp, observed_s)
}
\arguments{
\item{task_means_lst}{List of task mean matrices by group}

\item{permsamp}{Permutation sample matrix}

\item{observed_s}{Observed singular values}
}
\value{
Vector of counts
}
\description{
More efficient version that pre-computes task means and only shuffles.
}
\keyword{internal}
