% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_discovery.R
\name{pipeline_discover}
\alias{pipeline_discover}
\title{Discover BIDS Inputs for \code{plscli}}
\usage{
pipeline_discover(spec)
}
\arguments{
\item{spec}{Path to a YAML file or a pipeline specification list.}
}
\value{
The discovery manifest as a data.frame.
}
\description{
Discover BIDS Inputs for \code{plscli}
}
