% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_firstlevel.R
\name{pipeline_firstlevel_plan}
\alias{pipeline_firstlevel_plan}
\title{Plan First-Level Work Units}
\usage{
pipeline_firstlevel_plan(spec)
}
\arguments{
\item{spec}{Path to a YAML file or a pipeline specification list.}
}
\value{
The work-plan data.frame.
}
\description{
Plan First-Level Work Units
}
