% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_pls.R
\name{pipeline_pls_run}
\alias{pipeline_pls_run}
\title{Run PLS from a Planned Manifest}
\usage{
pipeline_pls_run(spec)
}
\arguments{
\item{spec}{Path to a YAML file or a pipeline specification list.}
}
\value{
A \code{pls_result} object.
}
\description{
Run PLS from a Planned Manifest
}
