% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_pls.R
\name{plot_asca_alignment}
\alias{plot_asca_alignment}
\title{Plot ASCA-PLS Partial-to-Effect Alignment}
\usage{
plot_asca_alignment(x, term = NULL, ...)
}
\arguments{
\item{x}{An \code{asca_pls_result}.}

\item{term}{Optional term label. When \code{NULL}, plot all terms.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot ASCA-PLS Partial-to-Effect Alignment
}
