% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_pls.R
\name{plot_asca_selection}
\alias{plot_asca_selection}
\title{Plot ASCA-PLS Model Selection Table}
\usage{
plot_asca_selection(x, p_column = c("p_adjusted", "p_value"), ...)
}
\arguments{
\item{x}{An \code{asca_pls_result}.}

\item{p_column}{\code{"p_adjusted"} or \code{"p_value"}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot ASCA-PLS Model Selection Table
}
