% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadings.R
\name{plot_pls_summary}
\alias{plot_pls_summary}
\title{Plot PLS Summary}
\usage{
plot_pls_summary(x, lv = 1)
}
\arguments{
\item{x}{A \code{pls_result} object}

\item{lv}{Latent variable for detailed plots}
}
\value{
A patchwork object
}
\description{
Creates a multi-panel summary plot of PLS results.
}
