% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_covariance.R
\name{pls-covariance}
\alias{pls-covariance}
\title{Covariance/Correlation Matrix Construction}
\description{
Constructs the cross-block covariance or correlation matrices used in PLS.
Ported from MATLAB rri_get_covcor.m
}
