% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_meancentering.R
\name{pls-meancentering}
\alias{pls-meancentering}
\title{Mean-Centering Operations for PLS}
\description{
Functions for computing task means and applying mean-centering.
Ported from MATLAB rri_task_mean.m and related functions.
}
