% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_perm.R
\name{pls-permutation}
\alias{pls-permutation}
\title{Permutation Testing for PLS}
\description{
Functions for generating permutation orders and running permutation tests.
Ported from MATLAB rri_perm_order.m
}
