% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_splithalf.R
\name{pls-splithalf}
\alias{pls-splithalf}
\title{Split-Half Validation for PLS}
\description{
Functions for split-half cross-validation to assess reliability of PLS results.
Based on Natasha's split-half routine.
}
