% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_scores.R
\name{pls_behav_scores}
\alias{pls_behav_scores}
\title{Get Behavior Scores}
\usage{
pls_behav_scores(
  stacked_behavdata,
  lvcorrs,
  num_groups,
  num_subj_lst,
  num_cond,
  n_lv
)
}
\arguments{
\item{stacked_behavdata}{Stacked behavior data}

\item{lvcorrs}{Latent variable correlations}

\item{num_groups}{Number of groups}

\item{num_subj_lst}{Subjects per group}

\item{num_cond}{Number of conditions}

\item{n_lv}{Number of latent variables}
}
\value{
Behavior scores matrix
}
\description{
Computes behavior scores for behavior and multiblock PLS.
These represent how strongly each behavior measure relates to each LV.
}
