% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_boot.R
\name{pls_boot_order_nonstrat}
\alias{pls_boot_order_nonstrat}
\title{Non-Stratified Bootstrap Orders}
\usage{
pls_boot_order_nonstrat(num_subj_lst, num_cond, num_boot)
}
\description{
Non-Stratified Bootstrap Orders
}
\keyword{internal}
