% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{pls_colors}
\alias{pls_colors}
\title{PLS Color Palettes}
\usage{
pls_colors(n = 5, palette = "diverging", reverse = FALSE)
}
\arguments{
\item{n}{Number of colors needed}

\item{palette}{Palette name: "diverging", "sequential", "qualitative"}

\item{reverse}{Logical, reverse palette order}
}
\value{
Character vector of hex colors
}
\description{
Color palettes designed for PLS visualizations.
}
\examples{
pls_colors(5, "diverging")
}
