% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_covariance.R
\name{pls_corr_maps_notall}
\alias{pls_corr_maps_notall}
\title{Correlation Maps for Subset of Conditions}
\usage{
pls_corr_maps_notall(behavdata, datamat, n_subj, bscan, cormode = 0L)
}
\arguments{
\item{behavdata}{Behavior matrix}

\item{datamat}{Brain data matrix}

\item{n_subj}{Number of subjects}

\item{bscan}{Vector of condition indices to use}

\item{cormode}{Correlation mode}
}
\value{
Correlation matrix for selected conditions
}
\description{
Computes correlation maps for a subset of conditions (used in multiblock PLS).
}
