% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_scores.R
\name{pls_design_scores}
\alias{pls_design_scores}
\title{Compute Design Scores}
\usage{
pls_design_scores(design, v)
}
\arguments{
\item{design}{Design matrix}

\item{v}{Design loadings from SVD (v matrix)}
}
\value{
Design scores matrix
}
\description{
Computes design scores by projecting the design matrix onto the design
loadings (v from SVD).
}
