% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_splithalf.R
\name{pls_get_covcor_subset}
\alias{pls_get_covcor_subset}
\title{Get Covariance for Subset of Rows}
\usage{
pls_get_covcor_subset(
  stacked_datamat,
  stacked_behavdata,
  num_groups,
  num_subj_lst,
  num_cond,
  method,
  row_idx,
  bscan,
  meancentering_type,
  cormode
)
}
\description{
Helper function to compute covariance/correlation matrices for a subset of
observations. This is mainly used by resampling routines.
}
\keyword{internal}
