% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_scores.R
\name{pls_lv_corrs}
\alias{pls_lv_corrs}
\title{Compute LV Correlations}
\usage{
pls_lv_corrs(
  stacked_behavdata,
  brain_scores,
  num_groups,
  num_subj_lst,
  num_cond,
  cormode = 0L
)
}
\arguments{
\item{stacked_behavdata}{Behavior data matrix}

\item{brain_scores}{Brain scores matrix}

\item{num_groups}{Number of groups}

\item{num_subj_lst}{Subjects per group}

\item{num_cond}{Number of conditions}

\item{cormode}{Correlation mode}
}
\value{
LV correlations matrix
}
\description{
Computes correlations between behavior data and brain scores.
These are the behavior loadings for behavior PLS.
}
