% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_scores.R
\name{pls_split_multiblock_v}
\alias{pls_split_multiblock_v}
\title{Extract Task/Behavior Scores Separately (Multiblock)}
\usage{
pls_split_multiblock_v(v, num_groups, num_cond, n_behav, bscan)
}
\arguments{
\item{v}{Design loadings from SVD}

\item{num_groups}{Number of groups}

\item{num_cond}{Number of conditions}

\item{n_behav}{Number of behavior measures}

\item{bscan}{Conditions used in behavior block}
}
\value{
List with task_v, behav_v components
}
\description{
For multiblock PLS, extracts the task block and behavior block components
of the design loadings (v) and scores separately.
}
