% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_svd.R
\name{pls_svd}
\alias{pls_svd}
\title{Standard PLS SVD}
\usage{
pls_svd(X, handle_missing = TRUE)
}
\arguments{
\item{X}{Matrix to decompose}

\item{handle_missing}{Logical; use misssvd if TRUE and missing values present}
}
\value{
List with u, d, v components
}
\description{
Wrapper for SVD used in PLS analysis. Uses economy SVD by default.
}
