% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{procrustes_cpp}
\alias{procrustes_cpp}
\title{Procrustes Rotation}
\usage{
procrustes_cpp(X, Y, scale = FALSE)
}
\arguments{
\item{X}{Target matrix}

\item{Y}{Source matrix to rotate}

\item{scale}{Allow scaling (default false)}
}
\value{
Rotated Y matrix
}
\description{
Aligns matrix Y to matrix X using Procrustes rotation.
Finds orthogonal matrix Q that minimizes ||X - Y*Q||_F
}
