% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progress_msg}
\alias{progress_msg}
\title{Print Progress Message}
\usage{
progress_msg(msg, progress = TRUE)
}
\arguments{
\item{msg}{Message to print}

\item{progress}{Logical, whether to print}
}
\description{
Prints a progress message if progress reporting is enabled.
}
\keyword{internal}
