% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_run.R
\name{quick_pls}
\alias{quick_pls}
\title{Quick Run for Simple Analyses}
\usage{
quick_pls(datamat_lst, num_subj_lst, num_cond, nperm = 1000, nboot = 500, ...)
}
\arguments{
\item{datamat_lst}{List of data matrices}

\item{num_subj_lst}{Subjects per group}

\item{num_cond}{Number of conditions}

\item{nperm}{Number of permutations}

\item{nboot}{Number of bootstrap samples}

\item{...}{Additional arguments to pls_analysis()}
}
\value{
A \code{pls_result} object
}
\description{
Convenience function for running simple task PLS without full builder API.
}
