% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_spec.R
\name{read_pipeline_spec}
\alias{read_pipeline_spec}
\title{Read a \code{plscli} Pipeline Specification}
\usage{
read_pipeline_spec(x)
}
\arguments{
\item{x}{Path to a YAML file or an already-loaded list.}
}
\value{
A normalized pipeline specification list.
}
\description{
Read a \code{plscli} Pipeline Specification
}
