% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mva_method.R
\name{register_method}
\alias{register_method}
\title{Register a Multivariate Analysis Method}
\usage{
register_method(name, method, aliases = character(0))
}
\arguments{
\item{name}{Canonical method name}

\item{method}{An \code{MvaMethod} R6 instance}

\item{aliases}{Character vector of alternative names}
}
\description{
Register a Multivariate Analysis Method
}
