% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_div}
\alias{safe_div}
\title{Safe Division}
\usage{
safe_div(num, denom)
}
\arguments{
\item{num}{Numerator}

\item{denom}{Denominator}
}
\value{
num/denom with 0/0 -> 0
}
\description{
Division that returns 0 for 0/0 cases.
}
\keyword{internal}
