% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{scale_color_pls_discrete}
\alias{scale_color_pls_discrete}
\title{PLS Qualitative Color Scale}
\usage{
scale_color_pls_discrete(...)
}
\arguments{
\item{...}{Arguments passed to scale_color_manual}
}
\value{
A ggplot2 scale object
}
\description{
ggplot2 color scale for discrete data (groups, conditions).
}
