% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{scale_fill_pls}
\alias{scale_fill_pls}
\title{PLS Diverging Fill Scale}
\usage{
scale_fill_pls(..., midpoint = 0, limits = NULL)
}
\arguments{
\item{...}{Arguments passed to scale_fill_gradientn}

\item{midpoint}{Midpoint of diverging scale (default 0)}

\item{limits}{Scale limits}
}
\value{
A ggplot2 scale object
}
\description{
ggplot2 fill scale for diverging data.
}
