% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{scale_fill_pls_discrete}
\alias{scale_fill_pls_discrete}
\title{PLS Qualitative Fill Scale}
\usage{
scale_fill_pls_discrete(...)
}
\arguments{
\item{...}{Arguments passed to scale_fill_manual}
}
\value{
A ggplot2 scale object
}
\description{
ggplot2 fill scale for discrete data.
}
