% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_pls.R
\name{selected_formula}
\alias{selected_formula}
\title{Extract the Selected ASCA-PLS Formula}
\usage{
selected_formula(x, ...)
}
\arguments{
\item{x}{An \code{asca_pls_result}.}

\item{...}{Unused.}
}
\value{
A one-sided formula.
}
\description{
Extract the Selected ASCA-PLS Formula
}
