% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_pls_seed}
\alias{set_pls_seed}
\title{Set Seed with Reproducibility}
\usage{
set_pls_seed(seed)
}
\arguments{
\item{seed}{Integer seed}
}
\description{
Sets random seed ensuring R and RNG stream are synchronized.
}
\keyword{internal}
