% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{singular_values}
\alias{singular_values}
\title{Extract Singular Values}
\usage{
singular_values(x, normalize = FALSE)
}
\arguments{
\item{x}{A \code{pls_result} object}

\item{normalize}{Logical, return variance explained (default FALSE)}
}
\value{
Vector of singular values or variance explained
}
\description{
Extracts singular values from a PLS result.
}
