% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stack_datamats}
\alias{stack_datamats}
\title{Stack Data Matrices}
\usage{
stack_datamats(datamat_lst)
}
\arguments{
\item{datamat_lst}{List of data matrices}
}
\value{
Stacked matrix
}
\description{
Stacks a list of data matrices vertically (by rows).
}
\keyword{internal}
