% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{task_mean_cpp}
\alias{task_mean_cpp}
\title{Fast Task Mean Computation}
\usage{
task_mean_cpp(datamat, n_subj, n_cond)
}
\arguments{
\item{datamat}{Data matrix (n_subj*n_cond x n_features)}

\item{n_subj}{Number of subjects}

\item{n_cond}{Number of conditions}
}
\value{
Task mean matrix (n_cond x n_features)
}
\description{
Computes task means from stacked data matrix.
}
