% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_spec.R
\name{validate_pipeline_spec}
\alias{validate_pipeline_spec}
\title{Validate a \code{plscli} Pipeline Specification}
\usage{
validate_pipeline_spec(x)
}
\arguments{
\item{x}{Path to a YAML file or a specification list.}
}
\value{
The validated specification with defaults applied.
}
\description{
Validate a \code{plscli} Pipeline Specification
}
