% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variance_explained}
\alias{variance_explained}
\title{Compute Explained Variance}
\usage{
variance_explained(s, cumulative = FALSE)
}
\arguments{
\item{s}{Vector of singular values}

\item{cumulative}{Logical, return cumulative variance}
}
\value{
Vector of variance explained
}
\description{
Computes percentage of variance explained by each singular value.
}
