% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_spec.R
\name{write_pipeline_template}
\alias{write_pipeline_template}
\title{Write an Example \code{plscli} YAML Specification}
\usage{
write_pipeline_template(path)
}
\arguments{
\item{path}{Output YAML path.}
}
\value{
The written path, invisibly.
}
\description{
Write an Example \code{plscli} YAML Specification
}
