% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xcor_cov_cpp}
\alias{xcor_cov_cpp}
\title{Fast Cross-Covariance}
\usage{
xcor_cov_cpp(design, datamat)
}
\arguments{
\item{design}{Design matrix (n_obs x n_design)}

\item{datamat}{Data matrix (n_obs x n_features)}
}
\value{
Cross-covariance matrix (n_design x n_features)
}
\description{
Computes covariance between design and data matrices.
}
\keyword{internal}
