% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xcor_dot_cpp}
\alias{xcor_dot_cpp}
\title{Fast Cross-Correlation (Dot Product)}
\usage{
xcor_dot_cpp(design, datamat)
}
\arguments{
\item{design}{Design matrix (n_obs x n_design)}

\item{datamat}{Data matrix (n_obs x n_features)}
}
\value{
Cross-product matrix (n_design x n_features)
}
\description{
Computes dot product between design and data matrices.
}
\keyword{internal}
