% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xcor_pearson_cpp}
\alias{xcor_pearson_cpp}
\title{Fast Cross-Correlation (Pearson)}
\usage{
xcor_pearson_cpp(design, datamat)
}
\arguments{
\item{design}{Design matrix (n_obs x n_design)}

\item{datamat}{Data matrix (n_obs x n_features)}
}
\value{
Cross-correlation matrix (n_design x n_features)
}
\description{
Computes Pearson correlation between design and data matrices.
}
\keyword{internal}
