% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_test_task_cpp}
\alias{perm_test_task_cpp}
\title{Fast Permutation Test for Task PLS}
\usage{
perm_test_task_cpp(
  stacked_datamat,
  permsamp,
  observed_s,
  num_groups,
  num_subj_lst,
  num_cond,
  meancentering_type
)
}
\arguments{
\item{stacked_datamat}{Stacked data matrix}

\item{permsamp}{Permutation sample matrix (total_rows x num_perm)}

\item{observed_s}{Observed singular values}

\item{num_groups}{Number of groups}

\item{num_subj_lst}{Subjects per group}

\item{num_cond}{Number of conditions}

\item{meancentering_type}{Mean-centering type}
}
\value{
Vector of counts (permuted s >= observed s)
}
\description{
Runs the inner loop of permutation testing for task PLS.
Computes singular values for each permutation and compares to observed.
}
\keyword{internal}
