% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plscli.R
\name{plscli_main}
\alias{plscli_main}
\title{Run the \code{plscli} Dispatcher}
\usage{
plscli_main(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{Character vector of command-line arguments.}
}
\value{
Invisibly returns an integer exit status.
}
\description{
Run the \code{plscli} Dispatcher
}
